refactor(plans): split plan/module detail helpers into focused modules
Extract generation alert/retry UI, module detail skeleton, timeline
state derivation, and shared task completion button without changing
behavior or public import paths.
