+++ src/app/(app)/plans/[id]/components/UpdateTaskStatusButton.tsx
@@ -1,10 +1,8 @@
 'use client';
 
-import { CheckCircle2, CircleDashed } from 'lucide-react';
 import type { JSX } from 'react';
-import { Button } from '@/components/ui/button';
-import { cn } from '@/lib/utils';
 import type { ProgressStatus } from '@/shared/types/db.types';
+import { TaskCompletionButton } from './TaskCompletionButton';
 
 interface UpdateTaskStatusButtonProps {
   taskId: string;
@@ -21,36 +19,13 @@ export const UpdateTaskStatusButton = (
   props: UpdateTaskStatusButtonProps,
 ): JSX.Element => {
   const { taskId, status, onStatusChange } = props;
-  const isCompleted = status === 'completed';
-  const buttonLabel = isCompleted ? 'Completed' : 'Mark Complete';
-
-  const handleClick = () => {
-    const nextStatus: ProgressStatus = isCompleted
-      ? 'not_started'
-      : 'completed';
-    onStatusChange(taskId, nextStatus);
-  };
 
   return (
-    <Button
-      onClick={handleClick}
-      aria-pressed={isCompleted}
-      aria-label={
-        isCompleted ? 'Mark task as incomplete' : 'Mark task as complete'
-      }
-      className={cn(
-        'rounded-xl border px-4 text-left',
-        isCompleted
-          ? 'border-success bg-success text-success-foreground hover:bg-success/90 dark:border-success dark:bg-success dark:hover:bg-success/90'
-          : 'border-border bg-secondary text-secondary-foreground hover:bg-accent hover:text-accent-foreground',
-      )}
-    >
-      {isCompleted ? (
-        <CheckCircle2 className="h-5 w-5" />
-      ) : (
-        <CircleDashed className="h-5 w-5" />
-      )}
-      {buttonLabel}
-    </Button>
+    <TaskCompletionButton
+      taskId={taskId}
+      status={status}
+      onStatusChange={onStatusChange}
+      variant="timeline"
+    />
   );
 };
