+++ src/app/(app)/plans/[id]/components/generation-alert-panel.tsx
@@ -1,0 +1,47 @@
+import { AlertCircle } from 'lucide-react';
+import type { ReactNode } from 'react';
+
+const ALERT_VARIANT_CLASSES = {
+  warning: {
+    container:
+      'flex items-start gap-3 rounded-lg border border-amber-500/20 bg-amber-500/10 p-4',
+    icon: 'mt-0.5 h-5 w-5 shrink-0 text-amber-600 dark:text-amber-400',
+    title: 'font-semibold text-amber-600 dark:text-amber-400',
+  },
+  destructive: {
+    container:
+      'flex items-start gap-3 rounded-lg border border-destructive/20 bg-destructive/10 p-4',
+    icon: 'mt-0.5 h-5 w-5 shrink-0 text-destructive',
+    title: 'font-semibold text-destructive',
+  },
+} as const;
+
+export function GenerationAlertPanel({
+  variant,
+  title,
+  body,
+  meta,
+  footer,
+}: {
+  variant: 'destructive' | 'warning';
+  title: string;
+  body: ReactNode;
+  meta?: ReactNode;
+  footer?: ReactNode;
+}) {
+  const classes = ALERT_VARIANT_CLASSES[variant];
+
+  return (
+    <div className={footer ? 'space-y-4' : undefined}>
+      <div className={classes.container}>
+        <AlertCircle className={classes.icon} />
+        <div className="space-y-1">
+          <p className={classes.title}>{title}</p>
+          <p className="text-sm text-muted-foreground">{body}</p>
+          {meta}
+        </div>
+      </div>
+      {footer}
+    </div>
+  );
+}
