+++ src/app/(app)/plans/[id]/modules/[moduleId]/components/TaskStatusButton.tsx
@@ -1,9 +1,7 @@
 'use client';
 
-import { CheckCircle2, Circle } from 'lucide-react';
 import type { JSX } from 'react';
-import { Button } from '@/components/ui/button';
-import { cn } from '@/lib/utils';
+import { TaskCompletionButton } from '@/app/(app)/plans/[id]/components/TaskCompletionButton';
 import type { ProgressStatus } from '@/shared/types/db.types';
 
 interface TaskStatusButtonProps {
@@ -23,36 +21,13 @@ export function TaskStatusButton({
   onStatusChange,
   disabled = false,
 }: TaskStatusButtonProps): JSX.Element {
-  const isCompleted = status === 'completed';
-
-  const handleClick = () => {
-    const nextStatus: ProgressStatus = isCompleted
-      ? 'not_started'
-      : 'completed';
-    onStatusChange(taskId, nextStatus);
-  };
-
   return (
-    <Button
-      onClick={handleClick}
+    <TaskCompletionButton
+      taskId={taskId}
+      status={status}
+      onStatusChange={onStatusChange}
       disabled={disabled}
-      aria-pressed={isCompleted}
-      aria-label={
-        isCompleted ? 'Mark task as incomplete' : 'Mark task as complete'
-      }
-      className={cn(
-        'rounded-xl px-4 transition-all',
-        isCompleted
-          ? 'bg-success text-success-foreground hover:bg-success/90'
-          : 'bg-white/50 text-foreground hover:bg-primary hover:text-primary-foreground dark:bg-card/50 dark:text-muted-foreground dark:hover:bg-primary dark:hover:text-primary-foreground',
-      )}
-    >
-      {isCompleted ? (
-        <CheckCircle2 className="h-5 w-5" />
-      ) : (
-        <Circle className="h-5 w-5" />
-      )}
-      {isCompleted ? 'Completed' : 'Mark Complete'}
-    </Button>
+      variant="lesson"
+    />
   );
 }
